Add basic zle integration
What this does now is if you source the zsh/smartfzf.sh
file (after building the c++, of course), whenever you
press ctrl+rightarrow it will show a list of files
matching the last token of the current input buffer
in the current directory.

Example:

    source zsh/smartfzf.sh
    rm -i clist<ctrl+rightarrow>

Then CMakeLists.txt will show up under the input line.
This is fairly useless because all it does it print
output, but it shows an example of things that we could
do in the future.

Right now, this is really only a proof-of-concept example,
but if we want to go any further, we'll have to clearly
plan out what we want the user interface to be, how much
of it should be handled by zsh-zle/bash-thing and how
much should be handled by our smartfzf program.